requirements.txt
environs~=8.0.0
python-dotenv
asyncpg
aiogram
gino